server/requirements.txt
Flask==1.0.2
tensorflow==1.10
gunicorn==19.9.0
numpy==1.14.5
opencv-python
Pillow